Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
 int main() {
-    int x,y,z;
+    double x,y,z;
     scanf("%lf %lf %lf", &x,&y,&z);
     printf("Average: %.2lf", ((x+y+z)/3));
     return 0;
